webapp/packages/api/user-service/requirements.txt
fastapi[all]
uvicorn[standard]
boto3  # For S3/MinIO
python-dotenv
litellm>=1.77.5
fastmcp>=2.12.4
google-auth
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
httpx>=0.25.0
faker>=22.0.0
PyYAML
CouchDB>=1.2
gunicorn
firebase-admin>=6.0.0
firebase-functions>=0.2.0
google-cloud-storage
cffi
a2wsgi
google-cloud-logging
ldap3>=2.9
